+++ print_hex.c
@@ -6,13 +6,13 @@
 /*   By: magebreh <magebreh@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/10 12:45:46 by magebreh          #+#    #+#             */
-/*   Updated: 2025/05/10 17:35:19 by magebreh         ###   ########.fr       */
+/*   Updated: 2025/05/12 11:13:49 by magebreh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-static int	count_digits(unsigned long n)
+static int	count_digits(uintptr_t n)
 {
 	int	count;
 
@@ -25,7 +25,7 @@ static int	count_digits(unsigned long n)
 	return (count);
 }
 
-char	*ft_utoa_hex(unsigned long n, int is_uppercase)
+char	*ft_utoa_hex(uintptr_t n, int is_uppercase)
 {
 	char			*res;
 	unsigned int	len;
@@ -56,7 +56,7 @@ char	*ft_utoa_hex(unsigned long n, int is_uppercase)
 static int	print_hex(va_list *args, int is_uppercase)
 {
 	int				i;
-	unsigned long	n;
+	uintptr_t	n;
 	char			*buff;
 
 	i = 0;
